Add always prefixed strategy
