+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/package.json
@@ -14,12 +14,12 @@
         "@inlang/paraglide-js-adapter-vite": "workspace:*",
         "dedent": "1.5.1",
         "magic-string": "^0.30.5",
-        "svelte": "^4.2.8",
-        "tiny-glob": "^0.2.9"
+        "svelte": "^4.2.8"
     },
     "devDependencies": {
         "typescript": "^5.3.2",
         "vite": "^5.0.4",
-        "vitest": "^0.34.6"
+        "vitest": "^0.34.6",
+        "tiny-glob": "^0.2.9"
     }
 }
