+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/src/plugin/index.ts
@@ -9,6 +9,7 @@ import {
 } from "../constants.js"
 import { preprocess } from "../index.js"
 import dedent from "dedent"
+import { getTranslatePathModuleCode } from "./translatePath.js"
 
 type UserConfig = Parameters<typeof vitePluginParaglide>[0]
 
@@ -58,54 +59,17 @@ function adapterSvelteKit(userConfig: UserConfig): Plugin {
 	}
 }
 
-function getTranslatePathModuleCode(): string {
-	return dedent`
-		import { sourceLanguageTag, availableLanguageTags } from "${OUTDIR_ALIAS}/runtime.js"
-
-		/**
-		 * Takes in a path without language information and
-		 * returns a path with language information.
-		 * 
-		 * @param {string} path
-		 * @param {string} lang
-		 * @returns {string}
-		 */
-		export function ${TRANSLATE_PATH_FUNCTION_NAME}(path, lang) {
-			// ignore external links & relative paths
-			if (!path.startsWith("/")) return path 
-
-			path = getPathWithoutLang(path)
-
-			//Don't prefix with the source language tag, that's the default
-			if (lang === sourceLanguageTag) return path
-
-			//Otherwise, prefix with the language tag
-			else return "/" + lang + path
-		}
-
-		/**
-		 * Removes the language tag from the path, if it exists.
-		 * @param {string} path
-		 */
-		function getPathWithoutLang(path) {
-			const [_, maybeLang, ...rest] = path.split("/")
-			if (availableLanguageTags.includes(maybeLang)) return "/" + rest.join("/")
-			else return path
-		}
-		`
-}
-
 function getHeaderComponentCode(): string {
 	return dedent`
 		<script>
 			import { availableLanguageTags } from "${OUTDIR_ALIAS}/runtime.js"
 			import { ${TRANSLATE_PATH_FUNCTION_NAME} } from "${TRANSLATE_PATH_MODULE_ID}"
-			// import { page } from "$app/stores"
+			import { page } from "$app/stores"
 		</script>
 
 		<svelte:head>
 			{#each availableLanguageTags as lang}
-				<link rel="alternate" hreflang={lang} href={${TRANSLATE_PATH_FUNCTION_NAME}("/", lang)} />
+				<link rel="alternate" hreflang={lang} href={${TRANSLATE_PATH_FUNCTION_NAME}($page.url.pathname, lang)} />
 			{/each}
 		</svelte:head>
 	`
